put isOwner / isLogged check for details page /setup a authguard

+++ src/app/shoe/shoe-details-page/shoe-details-page.component.ts
@@ -3,6 +3,7 @@ import { ShoesService } from '../shoes.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Shoe } from 'src/interfaces/shoe';
 import { UserService } from 'src/app/user/user.service';
+import { SharedService } from 'src/app/shared/shared.service';
 
 @Component({
   selector: 'app-shoe-details-page',
@@ -11,16 +12,18 @@ import { UserService } from 'src/app/user/user.service';
 })
 export class ShoeDetailsPageComponent {
   shoe: Shoe | undefined
-  
 
-  constructor(private shoeService: ShoesService, private activedRoute: ActivatedRoute, private router: Router,private userService:UserService) {
+
+  constructor(private shoeService: ShoesService, private activedRoute: ActivatedRoute, private router: Router, private sharedService: SharedService) {
 
   }
- 
-// get isOwner(){
-//   return this.userService.user$.
-// }
 
+  get isOwner() {
+    return this.sharedService.isOwner
+  }
+  get isLogged() {
+    return this.sharedService.isLogged
+  }
 
   ngOnInit(): void {
     this.getOneShoe()
@@ -34,10 +37,14 @@ export class ShoeDetailsPageComponent {
     })
   }
   removeShoe() {
-    const id = this.activedRoute.snapshot.params['shoeId']
-    this.shoeService.removeShoe(id).subscribe(() => {
-      this.router.navigate(['/catalog'])
+    if (confirm("are u sure>")) {
+      const id = this.activedRoute.snapshot.params['shoeId']
+      this.shoeService.removeShoe(id).subscribe(() => {
+        this.router.navigate(['/catalog'])
 
-    })
+      })
+    }
+
+
   }
 }
